libs/community/tests/integration_tests/navigation/requirements.txt
python-dotenv
nest_asyncio
pydantic
playwright
langchain
langchain-core
langchain-google-genai
langchain-ollama
langchain-community
langgraph
beautifulsoup4
lxml